backend/requirements.txt
# ── API
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
python-multipart==0.0.9

# ── Base de données
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pymongo==4.6.1
redis==5.0.1
clickhouse-connect==0.7.0

# ── Celery (tâches asynchrones)
celery==5.3.6
flower==2.0.1

# ── Data
pandas==2.1.4
numpy==1.26.3
dvc==3.38.1

# ── MinIO (Data Lake)
minio==7.2.3
boto3==1.34.0

# ── IA / LLM / RAG
langchain
langchain-community
langchain-core
langgraph
ollama
qdrant-client

# ── Embeddings CPU uniquement (sans PyTorch/CUDA)
# sentence-transformers retiré car trop lourd (télécharge PyTorch ~2GB)
# On utilise fastembed qui est léger et CPU-friendly
fastembed==0.3.1

# ── ML / Détection anomalies
scikit-learn==1.4.0
prophet==1.1.5
shap==0.44.0

# ── MLOps
mlflow==2.16.2
evidently==0.4.33
great-expectations==0.18.9

# ── Monitoring
prometheus-fastapi-instrumentator==7.0.0

# ── Utilitaires
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.27.0